lesson6/requirements.txt
aiofiles
alembic
asyncpg
python-dotenv
sqlalchemy